@@ -18,8 +18,8 @@
     "test:files": "npx jest --testPathIgnorePatterns js/output --testPathIgnorePatterns js/analyse",
     "test:output": "npx jest --testPathPattern js/output",
     "test:analytics": "npx rimraf analysis; npx jest --testPathPattern js/analyse; npm run test:allrules",
-    "test:allrules": "echo \"import AnalyticsTest from './js/analytics/analytics_test'; AnalyticsTest.outputAllRules()\" | node -r esm",
-    "analyse:tries": "echo \"import AnalyticsTrie from './js/analytics/analytics_trie'; AnalyticsTrie.output()\" | node -r esm",
+    "test:allrules": "echo \"import('./js/module_loader.js').then((m) => m.loadPromise('at')).then(m => m.default.outputAllRules())\" | node ",
+    "analyse:tries": "echo \"import('./js/module_loader.js').then((m) => m.loadPromise('atrie')).then(m => m.default.output())\" | node ",
     "clean": "npx rimraf dist js ts/output ts/json ts/analyse ts/actions coverage"
   },
   "maintainers": [
@@ -42,40 +42,42 @@
   ],
   "homepage": "https://github.com/speech-rule-engine/sre-tests",
   "dependencies": {
-    "@google-cloud/storage": "6.9.1",
-    "eslint-config-prettier": "^8.6.0",
+    "@google-cloud/storage": "6.10.1",
+    "@xmldom/xmldom": "^0.9.0-beta.8",
+    "eslint-config-prettier": "^8.8.0",
     "esm": "3.2.25",
-    "firebase-admin": "^11.5.0",
-    "firebase-tools": "^11.21.0",
-    "jest": "29.4.1",
+    "firebase-admin": "^11.8.0",
+    "firebase-tools": "^12.2.1",
+    "jest": "29.5.0",
     "mathjax-full": "3.2.2",
-    "npm-check-updates": "16.6.3",
+    "npm-check-updates": "16.10.12",
     "sre-visualiser": "^0.5.0",
-    "wicked-good-xpath": "1.3.0",
-    "xmldom-sre": "0.1.31"
+    "wicked-good-xpath": "1.3.0"
   },
   "devDependencies": {
-    "@babel/preset-env": "7.20.2",
-    "@types/jest": "29.4.0",
-    "@types/node": "18.11.18",
-    "@typescript-eslint/eslint-plugin": "5.49.0",
-    "@typescript-eslint/parser": "5.49.0",
-    "babel-jest": "29.4.1",
+    "@babel/preset-env": "7.22.4",
+    "@types/jest": "29.5.1",
+    "@types/node": "20.2.5",
+    "@typescript-eslint/eslint-plugin": "5.59.8",
+    "@typescript-eslint/parser": "5.59.8",
+    "babel-jest": "29.5.0",
     "babel-plugin-transform-class-properties": "6.24.1",
-    "eslint": "8.32.0",
+    "eslint": "8.41.0",
     "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-jsdoc": "39.6.9",
-    "prettier": "^2.8.3",
-    "terser-webpack-plugin": "5.3.6",
-    "ts-loader": "9.4.2",
-    "typescript": "4.9.4",
-    "webpack": "5.75.0",
-    "webpack-cli": "5.0.1"
+    "eslint-plugin-jsdoc": "45.0.0",
+    "prettier": "^2.8.8",
+    "terser-webpack-plugin": "5.3.9",
+    "ts-loader": "9.4.3",
+    "typescript": "5.0.4",
+    "typescript-tools": "0.3.1",
+    "webpack": "5.84.1",
+    "webpack-cli": "5.1.1"
   },
   "files": [
     "LICENSE",
     "README.md",
     "lib/"
   ],
-  "main": "dist/sretests.js"
+  "main": "dist/sretests.js",
+  "type": "module"
 }
